feat: add redux
