+++ src/index.js
@@ -1,5 +1,5 @@
 import * as serviceWorker from './serviceWorker';
-import store from './redux/state';
+import store from './redux/store';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
