+++ src/redux/redux_store.js
@@ -1,0 +1,3 @@
+const { createStore } = require("redux");
+
+let store = createStore ();
