+++ src/redux/sidebar_reducer.js
@@ -1,0 +1,6 @@
+const sidebar_reducer = (state, action) => {
+
+    return state
+}
+
+export default sidebar_reducer;
